requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
matplotlib
seaborn
joblib
fastapi
uvicorn
pyyaml

